Add files via upload
This was an assignment given to Year 2 Nanyang Business School Accoutancy students which I decided to try out. Completed this code within a few hours and was enough to achieve the Basic Functionality given by the assignment. The code reads from a text file and updates the output of the python program accordingly and also reflects input of the python program onto the text file.